Captioning embedded images is now a separate file Not a protocol Also now runs in parallel
